requirements.txt
playwright
fastapi
uvicorn[standard]
psycopg2-binary